Removed default controller from default route,
in order to match the default created by the Visual Studio project
template.

In case you're wondering why the "qux" URI segment was added, this is
because the 'incoming' request must match a route in the route table, and
the default expects a {controller} segment, where I've just now removed
the default value. If the're no matching route for the 'incoming' request,
the GetRouteData method returns null. Otherwise, the "qux" value isn't
used.