Added a Verify build configuration
and turned on Code Analysis. The purpose of having a third Verify build
configuration in addition to Debug and Release is that I don't want to run
Code Analysis for every debug build, as this is too slow and detrimental
to the Red/Green/Refactor TDD cycle. I also don't want to have Code
Analysis turned on in the Release build, as I don't want suppressions, if
any, to be visible in the final product (even though this is open source,
so this is strictly a subjective, aesthetic consideration).