Updated all NuGet packages,
mainly in order to update to the latest version of the ASP.NET Web API. I
hope this is the RTM version... It's a bit hard to tell from the NuGet
package itself.

After that I fixed a compiler error and made sure all tests are green. A
bit of subsequent refactoring now seems warranted.