Moved GetParameterValue to test project,
because, like other extension methods, this one shouldn't be public,
because it doesn't particularly relate to the main purpose of Hyprlinkr to
extend LINQ expressions. It should not be the responsibility of Hyprlinkr
to support extensions for LINQ expressions.

Additionally, this public method was only used by a test setup method, but
served no purpose related to Hyprlinkr.